Use extension methods to clean up PublicApi Program.cs file
