+++ src/PublicApi/Extensions/WebApplicationExtensions.cs
@@ -1,0 +1,35 @@
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.eShopWeb.Infrastructure.Data;
+using Microsoft.eShopWeb.Infrastructure.Identity;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using System;
+using System.Threading.Tasks;
+
+namespace Microsoft.eShopWeb.PublicApi.Extensions;
+
+public static class WebApplicationExtensions
+{
+    public static async Task SeedDatabaseAsync(this WebApplication app)
+    {
+        app.Logger.LogInformation("Seeding Database...");
+
+        using var scope = app.Services.CreateScope();
+        var scopedProvider = scope.ServiceProvider;
+        try
+        {
+            var catalogContext = scopedProvider.GetRequiredService<CatalogContext>();
+            await CatalogContextSeed.SeedAsync(catalogContext, app.Logger);
+
+            var userManager = scopedProvider.GetRequiredService<UserManager<ApplicationUser>>();
+            var roleManager = scopedProvider.GetRequiredService<RoleManager<IdentityRole>>();
+            var identityContext = scopedProvider.GetRequiredService<AppIdentityDbContext>();
+            await AppIdentityDbContextSeed.SeedAsync(identityContext, userManager, roleManager);
+        }
+        catch (Exception ex)
+        {
+            app.Logger.LogError(ex, "An error occurred seeding the DB.");
+        }
+    }
+}
